feat-508: refactor query extraction function
