+++ web/src/helpers/array.ts
@@ -28,6 +28,18 @@ export function takeFirstMaybe<T>(maybeArray: T | T[]): T | undefined {
   return undefined
 }
 
+export function joinAllMaybe(maybeArray: string | string[] | undefined): string | undefined {
+  if (!Array.isArray(maybeArray)) {
+    return maybeArray
+  }
+
+  if (maybeArray.length > 0) {
+    return maybeArray.join('/')
+  }
+
+  return undefined
+}
+
 /**
  * Checks if a string is present in an array of strings, case insensitive
  * (i.e. same as Array.includes() but the string comparison ignores case
