--- web/src/helpers/takeFirstMaybe.ts
@@ -1,28 +1,0 @@
-/**
- * Returns the first element of the array if non array is provided,
- * or `undefined` if empty array is provided, or passes through if
- * a non-array is provided.
- */
-export function takeFirstMaybe<T>(maybeArray: T | T[]): T | undefined {
-  if (!Array.isArray(maybeArray)) {
-    return maybeArray
-  }
-
-  if (maybeArray.length > 0) {
-    return maybeArray[0]
-  }
-
-  return undefined
-}
-
-export function joinAllMaybe(maybeArray: string | string[] | undefined): string | undefined {
-  if (!Array.isArray(maybeArray)) {
-    return maybeArray
-  }
-
-  if (maybeArray.length > 0) {
-    return maybeArray.join('/')
-  }
-
-  return undefined
-}
