+++ web/src/io/getQueryParamMaybe.ts
@@ -1,6 +1,5 @@
 import type { ParsedUrlQuery } from 'querystring'
-
-import { takeFirstMaybe } from 'src/helpers/takeFirstMaybe'
+import { takeFirstMaybe } from 'src/helpers/array'
 
 export function getQueryParamMaybe(urlQuery: ParsedUrlQuery, param: string): string | undefined {
   return takeFirstMaybe(urlQuery?.[param]) ?? undefined
